backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pdfplumber>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
groq>=0.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
lxml>=5.0.0
